+++ noodles-core/src/lib.rs
@@ -2,6 +2,7 @@
 
 //! **noodles-core** contains shared structures and behavior among noodles libraries.
 
+mod position;
 pub mod region;
 
-pub use self::region::Region;
+pub use self::{position::Position, region::Region};